Reorganised tests
This patch is mostly a reorganisation and extension of the tests.  Tests are
now run by going into the tests directory and running "make all".  Then also
"make clean" to clear up the files.

After running the QuickCheck test an execution profile is generated, with
a report in tests/Report.